services/ingestion-service/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "growthpilot-ingestion-service"
version = "0.1.0"
description = "Consumes ad.events from Redpanda, validates + normalizes, stores to Postgres."
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.6.1",
  "pydantic-settings>=2.2.1",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.6",
  "psycopg[binary]>=3.2.1",
  "aiokafka>=0.10.0",
  "python-dotenv>=1.0.1",
  "orjson>=3.10.0",

  # contracts
  "growthpilot-contracts>=0.1.0",

  # ✅ shared library (installed from inside the container)
  "growthpilot-shared @ file:///shared/python",
]

[tool.setuptools]
packages = ["app"]